pyproject.toml
[project]
name = "langfuse-poc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.10",
    "langgraph>=1.0.10",
    "langgraph-cli[inmem]>=0.1.0",
    "python-dotenv>=1.0.0",
    "langchain-openai>=1.1.11",
    "pydantic-settings>=2.13.1",
    "langfuse>=3.0.0",
    "fastapi>=0.135.1",
    "uvicorn>=0.41.0",
    "httpx>=0.28.1",
    "sse-starlette>=2.0.0",
]

[project.optional-dependencies]
# Install when Cloud SQL is provisioned:
#   uv sync --extra cloudsql
cloudsql = [
    "langgraph-checkpoint-postgres>=2.0.0",
    "psycopg[binary]>=3.0.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
]
